Generating (almost) all of the .NET primitive types is now possible via extension methods
A random number generator only has to be capable of generating a 32-bit word and it gets
the rest of the functionality for free.

+++ src/QuickCheck/Result.cs
@@ -9,12 +9,12 @@ namespace QuickCheck
     public class Result
     {
         private readonly int m_Tests;
-        private readonly int m_Seed;
+        private readonly ulong m_Seed;
         private readonly int m_Size;
         private readonly TestArgs m_Args;
         private readonly Exception m_Error;
 
-        private Result(int tests, int seed, int size, TestArgs args, Exception error)
+        private Result(int tests, ulong seed, int size, TestArgs args, Exception error)
         {
             m_Tests = tests;
             m_Seed = seed;
@@ -40,7 +40,7 @@ namespace QuickCheck
             return new Result(tests, 0, 0, null, null);
         }
 
-        public static Result Failure(int tests, int seed, int size, TestArgs args, Exception error)
+        public static Result Failure(int tests, ulong seed, int size, TestArgs args, Exception error)
         {
             return new Result(tests, seed, size, args, error);
         }
